backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
pytesseract==0.3.10
Werkzeug==2.3.7
requests==2.31.0
python-dateutil==2.8.2
hashlib-compat==1.0.1
logging==0.4.9.6
